+++ TennisWebapplication/Repositories/GenderRepository/IGenderRepository.cs
@@ -9,6 +9,6 @@ namespace TennisWebapplication.Repositories.GenderRepository
     public interface IGenderRepository
     {
         IEnumerable<Gender> GetAllGenders();
-        Gender GetGenderById(int id);
+        Gender GetGenderById(byte id);
     }
 }
